+++ app/services/jobs/workers/execution/worker.py
@@ -24,9 +24,11 @@ from qiskit_ibm_provider.utils import json_decoder
 from redis import Redis
 
 import settings
-from app.libs.quantum_executor.qiskit.executor import QiskitDynamicsExecutor, QiskitDynamics1QBSimulatorExecutor
+from app.libs.quantum_executor.qiskit.executor import (
+    QiskitDynamicsExecutor,
+    QiskitPulse1QExecutor,
+)
 from app.libs.quantum_executor.quantify.executor import QuantifyExecutor
-from app.libs.quantum_executor.qutip.executor import QuTipExecutor
 from app.libs.quantum_executor.utils.connections import get_executor_lock
 from app.libs.quantum_executor.utils.serialization import iqx_rld
 from app.utils.queues import QueuePool
@@ -44,10 +46,12 @@ STORAGE_ROOT = settings.STORAGE_ROOT
 BCC_MACHINE_ROOT_URL = settings.BCC_MACHINE_ROOT_URL
 DEFAULT_PREFIX = settings.DEFAULT_PREFIX
 
+
 # Redis connection
 # ----------------
 redis_connection = Redis()
 rq_queues = QueuePool(prefix=DEFAULT_PREFIX, connection=redis_connection)
+
 
 # Quantum Executor
 # ----------------
@@ -55,7 +59,7 @@ rq_queues = QueuePool(prefix=DEFAULT_PREFIX, connection=redis_connection)
 EXECUTOR_MAP = {
     "hardware": QuantifyExecutor,
     "qiskit": QiskitDynamicsExecutor,
-    "qutip": QuTipExecutor,
+    "qiskit_pulse_1q": QiskitPulse1QExecutor,
 }
 executor = EXECUTOR_MAP[settings.EXECUTOR_TYPE](
     config_file=settings.EXECUTOR_CONFIG_FILE