+++ settings.py
@@ -25,6 +25,9 @@ from starlette.datastructures import URL
 env_file = os.environ.get("ENV_FILE", default=".env")
 config = Config(Path(__file__).parent / env_file)
 
+# Automatic root directory settings
+APP_ROOT_DIR = Path(__file__).parent / "app"
+
 # Misc settings
 APP_SETTINGS = config("APP_SETTINGS", cast=str, default="production")
 IS_AUTH_ENABLED = config("IS_AUTH_ENABLED", cast=bool, default=True)
@@ -54,7 +57,9 @@ EXECUTOR_DATA_DIRNAME = config(
     "EXECUTOR_DATA_DIRNAME", cast=str, default="executor_data"
 )
 
-_executor_data_dir_path = os.path.join(STORAGE_ROOT, EXECUTOR_DATA_DIRNAME)
+_executor_data_dir_path = os.path.join(
+    STORAGE_ROOT, DEFAULT_PREFIX, EXECUTOR_DATA_DIRNAME
+)
 if not os.path.exists(_executor_data_dir_path):
     os.makedirs(_executor_data_dir_path)
 EXECUTOR_DATA_DIR = _executor_data_dir_path